ida_ground_control_station/requirements.txt
PyQt5==5.15.10
PyQtWebEngine==5.15.6
pyserial==3.5 